Updated the Community video to 2024 talk - "The Power of Community"
In looking ahead to RoboCon 2025 and in looking back over the years of
building community, i thought it would be a good time to update the
community talk that is presented on the home page. I have switched out
the previous "The Importance of Open Source Communities" [2GDrtvz_1Ds]
from 2018 with "The Power of Community" [qbFNSVSRBZ8] from 2024. I have
reached out to Guido, Merel, Manana, Yuri and Kelby and got all of their
OK for presenting this on the homepage.

Eventually I was thinking of putting a small carousel of three videos;
one where you can see the main one and notice their are two others to
choose from. But for now I am going very simple and just updating the
one.